Fix design flaws
Some class "A" should not be calling methods from some class "B" if "B" has a list of "A"s. This create cyclic dependencies that forces developers to make changes in at least both classes when a functionality is added, updated or removed. Besides it violates, among others , the "Hollywood Principle".
- Remove cyclic dependency between ReportsWindowViewModel and LoggedExperimentalUnitViewModel
- Remove cyclic dependency between ReportsWindowViewModel and LoggedForkViewModel
- Remove cyclic dependency between ReportsWindowViewModel and LoggedForkValueViewModel